Restructured the dir and added voronoi plotting.
