+++ requirements.txt
@@ -5,3 +5,4 @@ Pillow~=10.1.0
 gbox~=0.1.1
 h5py~=3.10.0
 tqdm~=4.66.1
+scipy~=1.11.4
